Fixed sorting order of canvases!!
The current weapon used to be hidden behind the portion of the currently unused weapon that it is overlapping with. This is not supposed to happen. I fixed it. Now, I have to adjust the position of the base anchor during the animations, because I forgot to factor in the aspect ratio.